requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-phonenumber-field==7.3.0
idna==3.7
phonenumberslite==8.13.35
requests==2.31.0
sqlparse==0.5.0
urllib3==2.2.1
